Reorganize sourcecode
