+++ .gitignore
@@ -1,3 +1,3 @@
-src/*.js
+src/**/*.js
 node_modules/
 .env
