+++ src/sticker.ts
